+++ src/forms/uGBPlayer.c
@@ -341,6 +341,7 @@ static void StartEmulation(void)
 	UInt16 vrn, latestPrefSize;
 	UInt32 mask;
 	struct UgbPrefs *prefs;
+	static int keyBits[] = {KEY_BIT_LEFT, KEY_BIT_UP, KEY_BIT_RIGHT, KEY_BIT_DOWN, KEY_BIT_SEL, KEY_BIT_START, KEY_BIT_B, KEY_BIT_A};
 
 	if (errNone == WinScreenGetAttribute(winScreenWidth, &screenPixelW) &&
 			errNone == WinScreenGetAttribute(winScreenHeight, &screenPixelH) &&
@@ -381,44 +382,10 @@ static void StartEmulation(void)
 
 				MemSet(pd->keyMapping, sizeof(pd->keyMapping), 0);
 
-				if (prefs->keys[0] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[0])] = KEY_BIT_LEFT;
-				}
-
-				if (prefs->keys[1] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[1])] = KEY_BIT_UP;
-				}
-
-				if (prefs->keys[2] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[2])] = KEY_BIT_RIGHT;
-				}
-
-				if (prefs->keys[3] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[3])] = KEY_BIT_DOWN;
-				}
-
-				if (prefs->keys[4] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[4])] = KEY_BIT_SEL;
-				}
-
-				if (prefs->keys[5] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[5])] = KEY_BIT_START;
-				}
-
-				if (prefs->keys[6] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[6])] = KEY_BIT_B;
-				}
-
-				if (prefs->keys[7] != 0)
-				{
-					pd->keyMapping[__builtin_ctzl(prefs->keys[7])] = KEY_BIT_A;
+				for (int i = 0; i < 8; ++i) {
+					if (prefs->keys[i] != 0) {
+						pd->keyMapping[__builtin_ctzl(prefs->keys[i])] = keyBits[i];
+					}
 				}
 
 				mask = KeySetMask(0);